refactore(backend): cleanup queue handler and description
