+++ backend/internal/queue/handler.go
@@ -1,962 +1,16 @@
 package queue
 
 import (
-	"bytes"
 	"context"
 	"encoding/json"
-	"errors"
 	"fmt"
-	"hash/fnv"
-	"path/filepath"
-	"slices"
-	"sort"
-	"strconv"
-	"strings"
-	"time"
 
 	"github.com/OFFIS-RIT/kiwi/backend/internal/db"
-	"github.com/OFFIS-RIT/kiwi/backend/internal/storage"
-	"github.com/OFFIS-RIT/kiwi/backend/internal/util"
 	"github.com/OFFIS-RIT/kiwi/backend/pkg/logger"
-	graphstorage "github.com/OFFIS-RIT/kiwi/backend/pkg/store/base"
 
-	"github.com/aws/aws-sdk-go-v2/aws"
-	awss3 "github.com/aws/aws-sdk-go-v2/service/s3"
-	"github.com/jackc/pgx/v5"
-	"github.com/jackc/pgx/v5/pgtype"
 	"github.com/jackc/pgx/v5/pgxpool"
-	"github.com/pkoukk/tiktoken-go"
 	"github.com/rabbitmq/amqp091-go"
-
-	"github.com/OFFIS-RIT/kiwi/backend/pkg/ai"
-	"github.com/OFFIS-RIT/kiwi/backend/pkg/graph"
-	"github.com/OFFIS-RIT/kiwi/backend/pkg/loader"
-	"github.com/OFFIS-RIT/kiwi/backend/pkg/loader/audio"
-	"github.com/OFFIS-RIT/kiwi/backend/pkg/loader/csv"
-	"github.com/OFFIS-RIT/kiwi/backend/pkg/loader/doc"
-	"github.com/OFFIS-RIT/kiwi/backend/pkg/loader/excel"
-	"github.com/OFFIS-RIT/kiwi/backend/pkg/loader/image"
-	"github.com/OFFIS-RIT/kiwi/backend/pkg/loader/ocr"
-	"github.com/OFFIS-RIT/kiwi/backend/pkg/loader/pdf"
-	"github.com/OFFIS-RIT/kiwi/backend/pkg/loader/pptx"
-	"github.com/OFFIS-RIT/kiwi/backend/pkg/loader/s3"
 )
-
-type QueueProjectFileMsg struct {
-	Message       string            `json:"message"`
-	ProjectID     int64             `json:"project_id"`
-	CorrelationID string            `json:"correlation_id,omitempty"`
-	BatchID       int               `json:"batch_id,omitempty"`
-	TotalBatches  int               `json:"total_batches,omitempty"`
-	ProjectFiles  *[]db.ProjectFile `json:"project_files,omitempty"`
-	Operation     string            `json:"operation,omitempty"`
-}
-
-type QueueDescriptionJobMsg struct {
-	ProjectID       int64   `json:"project_id"`
-	CorrelationID   string  `json:"correlation_id"`
-	JobID           int     `json:"job_id"`
-	TotalJobs       int     `json:"total_jobs"`
-	EntityIDs       []int64 `json:"entity_ids,omitempty"`
-	RelationshipIDs []int64 `json:"relationship_ids,omitempty"`
-}
-
-func ProcessGraphMessage(
-	ctx context.Context,
-	s3Client *awss3.Client,
-	aiClient ai.GraphAIClient,
-	ch *amqp091.Channel,
-	conn *pgxpool.Pool,
-	msg string,
-) error {
-	data := new(QueueProjectFileMsg)
-	err := json.Unmarshal([]byte(msg), &data)
-	if err != nil {
-		return err
-	}
-	projectId := data.ProjectID
-
-	q := db.New(conn)
-
-	if data.CorrelationID != "" {
-		_ = q.UpdateBatchStatus(ctx, db.UpdateBatchStatusParams{
-			CorrelationID: data.CorrelationID,
-			BatchID:       int32(data.BatchID),
-			Column3:       "extracting",
-		})
-	}
-
-	isUpdate := data.Operation == "update"
-	statType := "graph_creation"
-	projectState := "create"
-	if isUpdate {
-		statType = "graph_update"
-		projectState = "update"
-	}
-	if _, err := q.UpdateProjectState(ctx, db.UpdateProjectStateParams{
-		ID:    projectId,
-		State: projectState,
-	}); err != nil {
-		logger.Warn("[Queue] Failed to update project state at graph start", "project_id", projectId, "state", projectState, "err", err)
-	}
-
-	s3Bucket := util.GetEnvString("AWS_BUCKET", "kiwi")
-	s3L := s3.NewS3GraphFileLoaderWithClient(s3Bucket, s3Client)
-	files := make([]loader.GraphFile, 0)
-
-	for _, file := range *data.ProjectFiles {
-		metadataText := ""
-		if file.Metadata.Valid {
-			metadataText = file.Metadata.String
-		}
-
-		ext := filepath.Ext(file.FileKey)
-		ext = strings.ReplaceAll(ext, ".", "")
-		ext = strings.ToLower(ext)
-
-		switch ext {
-		case "xlsx", "xls":
-			baseName := strings.TrimSuffix(filepath.Base(file.FileKey), "."+ext)
-			dir := filepath.Dir(file.FileKey)
-			prefix := fmt.Sprintf("%s/%s_", dir, baseName)
-
-			sheetFiles, err := storage.ListFilesWithPrefix(ctx, s3Client, prefix)
-			if err != nil {
-				return err
-			}
-
-			sheetIndex := 0
-			for _, sheetFile := range sheetFiles {
-				if !strings.HasSuffix(sheetFile, ".txt") {
-					continue
-				}
-				f := loader.NewGraphDocumentFile(loader.NewGraphFileParams{
-					ID:        fmt.Sprintf("%d-sheet-%d", file.ID, sheetIndex),
-					FilePath:  sheetFile,
-					MaxTokens: 500,
-					Loader:    s3L,
-					Metadata:  metadataText,
-				})
-
-				files = append(files, f)
-				sheetIndex++
-			}
-		case "csv":
-			key := file.FileKey
-			base := filepath.Base(file.FileKey)
-			nameWithoutExt := strings.TrimSuffix(base, filepath.Ext(base))
-			name := fmt.Sprintf("%s.txt", nameWithoutExt)
-			key = fmt.Sprintf("%s/%s", filepath.Dir(file.FileKey), name)
-
-			f := loader.NewGraphCSVFile(loader.NewGraphFileParams{
-				ID:        fmt.Sprintf("%d", file.ID),
-				FilePath:  key,
-				MaxTokens: 500,
-				Loader:    s3L,
-				Metadata:  metadataText,
-			})
-			files = append(files, f)
-		case "jpg", "jpeg", "png", "bmp", "gif", "tiff", "heic", "webp":
-			key := file.FileKey
-			base := filepath.Base(file.FileKey)
-			nameWithoutExt := strings.TrimSuffix(base, filepath.Ext(base))
-			name := fmt.Sprintf("%s.txt", nameWithoutExt)
-			key = fmt.Sprintf("%s/%s", filepath.Dir(file.FileKey), name)
-
-			f := loader.NewGraphImageFile(loader.NewGraphFileParams{
-				ID:        fmt.Sprintf("%d", file.ID),
-				FilePath:  key,
-				MaxTokens: 500,
-				Loader:    s3L,
-				Metadata:  metadataText,
-			})
-			files = append(files, f)
-		case "mp3", "wav", "mpeg", "mpga", "m4a", "ogg", "webm":
-			key := file.FileKey
-			base := filepath.Base(file.FileKey)
-			nameWithoutExt := strings.TrimSuffix(base, filepath.Ext(base))
-			name := fmt.Sprintf("%s.txt", nameWithoutExt)
-			key = fmt.Sprintf("%s/%s", filepath.Dir(file.FileKey), name)
-
-			f := loader.NewGraphAudioFile(loader.NewGraphFileParams{
-				ID:        fmt.Sprintf("%d", file.ID),
-				FilePath:  key,
-				MaxTokens: 500,
-				Loader:    s3L,
-				Metadata:  metadataText,
-			})
-			files = append(files, f)
-		case "doc", "docx", "odt", "pptx", "pdf":
-			key := file.FileKey
-			base := filepath.Base(file.FileKey)
-			nameWithoutExt := strings.TrimSuffix(base, filepath.Ext(base))
-			name := fmt.Sprintf("%s.txt", nameWithoutExt)
-			key = fmt.Sprintf("%s/%s", filepath.Dir(file.FileKey), name)
-
-			f := loader.NewGraphDocumentFile(loader.NewGraphFileParams{
-				ID:        fmt.Sprintf("%d", file.ID),
-				FilePath:  key,
-				MaxTokens: 500,
-				Loader:    s3L,
-				Metadata:  metadataText,
-			})
-			files = append(files, f)
-		case "txt", "md":
-			f := loader.NewGraphDocumentFile(loader.NewGraphFileParams{
-				ID:        fmt.Sprintf("%d", file.ID),
-				FilePath:  file.FileKey,
-				MaxTokens: 500,
-				Loader:    s3L,
-				Metadata:  metadataText,
-			})
-			files = append(files, f)
-		default:
-			key := file.FileKey
-			base := filepath.Base(file.FileKey)
-			nameWithoutExt := strings.TrimSuffix(base, filepath.Ext(base))
-			name := fmt.Sprintf("%s.txt", nameWithoutExt)
-			key = fmt.Sprintf("%s/%s", filepath.Dir(file.FileKey), name)
-
-			f := loader.NewGraphGenericFile(
-				loader.NewGraphFileParams{
-					ID:        fmt.Sprintf("%d", file.ID),
-					FilePath:  key,
-					MaxTokens: 500,
-					Loader:    s3L,
-				},
-				file.Name,
-			)
-			files = append(files, f)
-		}
-	}
-
-	tokenCount := 0
-	for _, f := range files {
-		fileID := f.ID
-		if idx := strings.Index(fileID, "-sheet-"); idx != -1 {
-			fileID = fileID[:idx]
-		}
-		id, err := strconv.ParseInt(fileID, 10, 64)
-		if err != nil {
-			return err
-		}
-
-		tokens, err := q.GetTokenCountOfFile(ctx, id)
-		if err != nil {
-			return err
-		}
-		tokenCount += int(tokens)
-	}
-
-	prediction, err := q.PredictProjectProcessTime(ctx, db.PredictProjectProcessTimeParams{
-		Duration: int64(tokenCount),
-		StatType: statType,
-	})
-	if err != nil {
-		prediction = 0
-	}
-	logger.Info("[Queue] Prediction for graph operation", "operation", data.Operation, "tokens", tokenCount, "time_ms", prediction)
-
-	if data.CorrelationID != "" {
-		_ = q.UpdateBatchEstimatedDuration(ctx, db.UpdateBatchEstimatedDurationParams{
-			CorrelationID:     data.CorrelationID,
-			BatchID:           int32(data.BatchID),
-			EstimatedDuration: prediction,
-		})
-	}
-
-	graphClient, err := graph.NewGraphClient(graph.NewGraphClientParams{
-		TokenEncoder:  "o200k_base",
-		ParallelFiles: 1,
-	})
-	if err != nil {
-		return err
-	}
-	storageClient, err := graphstorage.NewGraphDBStorageWithConnection(ctx, conn, aiClient, []string{})
-	if err != nil {
-		return err
-	}
-	deleteStagedData := func() error {
-		if data.CorrelationID == "" {
-			return nil
-		}
-		return storageClient.DeleteStagedData(ctx, data.CorrelationID, data.BatchID)
-	}
-	if err := deleteStagedData(); err != nil {
-		return fmt.Errorf("failed to cleanup staged data before extraction: %w", err)
-	}
-
-	graphID := fmt.Sprintf("%d", projectId)
-	start := time.Now()
-
-	logger.Debug("[Queue] Starting extraction phase", "project_id", projectId, "batch_id", data.BatchID)
-	err = graphClient.ExtractAndStage(ctx, files, graphID, aiClient, storageClient, data.CorrelationID, data.BatchID)
-	if err != nil {
-		if data.CorrelationID != "" {
-			_ = q.UpdateBatchStatus(ctx, db.UpdateBatchStatusParams{
-				CorrelationID: data.CorrelationID,
-				BatchID:       int32(data.BatchID),
-				Column3:       "failed",
-				ErrorMessage:  pgtype.Text{String: err.Error(), Valid: true},
-			})
-		}
-		if cleanupErr := deleteStagedData(); cleanupErr != nil {
-			logger.Warn("[Queue] Failed to delete staged data", "correlation_id", data.CorrelationID, "batch_id", data.BatchID, "err", cleanupErr)
-		}
-		return err
-	}
-
-	logger.Debug("[Queue] Acquiring advisory lock", "project_id", projectId)
-	err = q.AcquireProjectLock(ctx, projectId)
-	if err != nil {
-		if cleanupErr := deleteStagedData(); cleanupErr != nil {
-			logger.Warn("[Queue] Failed to delete staged data", "correlation_id", data.CorrelationID, "batch_id", data.BatchID, "err", cleanupErr)
-		}
-		return fmt.Errorf("failed to acquire project lock: %w", err)
-	}
-	defer func() {
-		if unlockErr := q.ReleaseProjectLock(ctx, projectId); unlockErr != nil {
-			errMsg := unlockErr.Error()
-			if strings.Contains(errMsg, "conn closed") {
-				logger.Debug("[Queue] Failed to release project lock (connection closed, lock auto-released)", "project_id", projectId, "err", unlockErr)
-			} else {
-				logger.Error("[Queue] Failed to release project lock", "project_id", projectId, "err", unlockErr)
-			}
-		}
-		logger.Debug("[Queue] Released advisory lock", "project_id", projectId)
-	}()
-
-	if data.CorrelationID != "" {
-		_ = q.UpdateBatchStatus(ctx, db.UpdateBatchStatusParams{
-			CorrelationID: data.CorrelationID,
-			BatchID:       int32(data.BatchID),
-			Column3:       "indexing",
-		})
-	}
-
-	_, err = q.UpdateProjectState(ctx, db.UpdateProjectStateParams{
-		ID:    projectId,
-		State: projectState,
-	})
-	if err != nil {
-		if cleanupErr := deleteStagedData(); cleanupErr != nil {
-			logger.Warn("[Queue] Failed to delete staged data", "correlation_id", data.CorrelationID, "batch_id", data.BatchID, "err", cleanupErr)
-		}
-		return err
-	}
-	err = graphClient.MergeFromStaging(ctx, files, graphID, aiClient, storageClient, data.CorrelationID, data.BatchID)
-	if err != nil {
-		if data.CorrelationID != "" {
-			_ = q.UpdateBatchStatus(ctx, db.UpdateBatchStatusParams{
-				CorrelationID: data.CorrelationID,
-				BatchID:       int32(data.BatchID),
-				Column3:       "failed",
-				ErrorMessage:  pgtype.Text{String: err.Error(), Valid: true},
-			})
-		}
-		if cleanupErr := deleteStagedData(); cleanupErr != nil {
-			logger.Warn("[Queue] Failed to delete staged data", "correlation_id", data.CorrelationID, "batch_id", data.BatchID, "err", cleanupErr)
-		}
-		return err
-	}
-
-	if cleanupErr := deleteStagedData(); cleanupErr != nil {
-		logger.Warn("[Queue] Failed to delete staged data", "correlation_id", data.CorrelationID, "batch_id", data.BatchID, "err", cleanupErr)
-	}
-
-	duration := time.Since(start)
-	q.AddProcessTime(ctx, db.AddProcessTimeParams{
-		ProjectID: projectId,
-		Amount:    int32(tokenCount),
-		Duration:  duration.Milliseconds(),
-		StatType:  statType,
-	})
-
-	if data.CorrelationID != "" {
-		_ = q.UpdateBatchStatus(ctx, db.UpdateBatchStatusParams{
-			CorrelationID: data.CorrelationID,
-			BatchID:       int32(data.BatchID),
-			Column3:       "completed",
-		})
-
-		allDone, checkErr := q.AreAllBatchesCompleted(ctx, data.CorrelationID)
-		if checkErr == nil && allDone {
-			_, err := KickoffDescriptionJobsForCorrelation(ctx, ch, conn, data.CorrelationID, projectId)
-			if err != nil {
-				return err
-			}
-
-			descDone, descErr := q.AreAllDescriptionJobsCompleted(ctx, data.CorrelationID)
-			if descErr != nil {
-				return descErr
-			}
-			if descDone {
-				latestCorrelationID, latestErr := q.GetLatestCorrelationForProject(ctx, projectId)
-				if latestErr != nil {
-					logger.Warn("[Queue] Failed to fetch latest correlation for project", "project_id", projectId, "correlation_id", data.CorrelationID, "err", latestErr)
-				} else if latestCorrelationID == data.CorrelationID {
-					if _, err := q.UpdateProjectState(ctx, db.UpdateProjectStateParams{ID: projectId, State: "ready"}); err != nil {
-						logger.Error("[Queue] Failed to set project state to ready", "project_id", projectId, "correlation_id", data.CorrelationID, "err", err)
-					}
-				}
-			}
-		}
-	} else {
-		if err := storageClient.GenerateDescriptions(ctx, files); err != nil {
-			logger.Error("[Queue] Failed to generate descriptions", "project_id", projectId, "err", err)
-		}
-		if _, err := q.UpdateProjectState(ctx, db.UpdateProjectStateParams{ID: projectId, State: "ready"}); err != nil {
-			logger.Error("[Queue] Failed to set project state to ready", "project_id", projectId, "err", err)
-		}
-	}
-
-	return nil
-}
-
-func ProcessDeleteMessage(
-	ctx context.Context,
-	s3Client *awss3.Client,
-	aiClient ai.GraphAIClient,
-	ch *amqp091.Channel,
-	conn *pgxpool.Pool,
-	msg string,
-) error {
-	type deleteJsonMsg struct {
-		ProjectID int64 `json:"project_id"`
-	}
-
-	data := new(deleteJsonMsg)
-	err := json.Unmarshal([]byte(msg), &data)
-	if err != nil {
-		return err
-	}
-	projectId := data.ProjectID
-	graphID := fmt.Sprintf("%d", projectId)
-
-	q := db.New(conn)
-
-	logger.Debug("[Queue] Acquiring advisory lock for delete", "project_id", projectId)
-	err = q.AcquireProjectLock(ctx, projectId)
-	if err != nil {
-		return fmt.Errorf("failed to acquire project lock: %w", err)
-	}
-	defer func() {
-		if unlockErr := q.ReleaseProjectLock(ctx, projectId); unlockErr != nil {
-			errMsg := unlockErr.Error()
-			if strings.Contains(errMsg, "conn closed") {
-				logger.Debug("[Queue] Failed to release project lock (connection closed, lock auto-released)", "project_id", projectId, "err", unlockErr)
-			} else {
-				logger.Error("[Queue] Failed to release project lock", "project_id", projectId, "err", unlockErr)
-			}
-		}
-		logger.Debug("[Queue] Released advisory lock", "project_id", projectId)
-	}()
-
-	deletedFiles, err := q.GetDeletedProjectFiles(ctx, projectId)
-	if err != nil {
-		return err
-	}
-	fileKeys := make([]string, 0, len(deletedFiles))
-	for _, file := range deletedFiles {
-		fileKeys = append(fileKeys, file.FileKey)
-	}
-
-	graphClient, err := graph.NewGraphClient(graph.NewGraphClientParams{
-		TokenEncoder:  "o200k_base",
-		ParallelFiles: 1,
-	})
-	if err != nil {
-		return err
-	}
-
-	storageClient, err := graphstorage.NewGraphDBStorageWithConnection(ctx, conn, aiClient, []string{})
-	if err != nil {
-		return err
-	}
-
-	_, err = q.UpdateProjectState(ctx, db.UpdateProjectStateParams{
-		ID:    projectId,
-		State: "update",
-	})
-	if err != nil {
-		return err
-	}
-	defer q.UpdateProjectState(ctx, db.UpdateProjectStateParams{
-		ID:    projectId,
-		State: "ready",
-	})
-
-	start := time.Now()
-	err = graphClient.DeleteGraph(ctx, graphID, aiClient, storageClient)
-	if err != nil {
-		return err
-	}
-	duration := time.Since(start)
-
-	logger.Info("[Queue] Delete and regenerate completed", "project_id", projectId, "duration_sec", duration.Seconds())
-
-	for _, fileKey := range fileKeys {
-		if err := storage.DeleteFile(ctx, s3Client, fileKey); err != nil {
-			logger.Warn("[Queue] Failed to delete S3 file", "file_key", fileKey, "err", err)
-		}
-	}
-
-	return nil
-}
-
-func ProcessPreprocess(
-	ctx context.Context,
-	s3Client *awss3.Client,
-	aiClient ai.GraphAIClient,
-	ch *amqp091.Channel,
-	conn *pgxpool.Pool,
-	msg string,
-) error {
-	var data QueueProjectFileMsg
-	err := json.Unmarshal([]byte(msg), &data)
-	if err != nil {
-		return err
-	}
-
-	q := db.New(conn)
-
-	projectState := "create"
-	if data.Operation == "update" {
-		projectState = "update"
-	}
-	if _, err := q.UpdateProjectState(ctx, db.UpdateProjectStateParams{ID: data.ProjectID, State: projectState}); err != nil {
-		logger.Warn("[Queue] Failed to update project state at preprocess start", "project_id", data.ProjectID, "state", projectState, "err", err)
-	}
-
-	if data.CorrelationID != "" {
-		_ = q.UpdateBatchStatus(ctx, db.UpdateBatchStatusParams{
-			CorrelationID: data.CorrelationID,
-			BatchID:       int32(data.BatchID),
-			Column3:       "preprocessing",
-		})
-	}
-
-	files := make([]loader.GraphFile, 0)
-	noProcessingFiles := make([]loader.GraphFile, 0)
-	pageCount := 0
-	s3Bucket := util.GetEnvString("AWS_BUCKET", "kiwi")
-	for _, upload := range *data.ProjectFiles {
-		metadataText := ""
-		if upload.Metadata.Valid {
-			metadataText = upload.Metadata.String
-		}
-
-		s3L := s3.NewS3GraphFileLoaderWithClient(s3Bucket, s3Client)
-		ocrL := ocr.NewOCRGraphLoader(ocr.NewOCRGraphLoaderParams{
-			AIClient: aiClient,
-		})
-
-		ext := filepath.Ext(upload.Name)
-		ext = strings.ReplaceAll(ext, ".", "")
-		ext = strings.ToLower(ext)
-
-		switch ext {
-		case "doc", "docx", "odt":
-			docL := doc.NewDocOcrGraphLoader(s3L, ocrL)
-			f := loader.NewGraphDocumentFile(loader.NewGraphFileParams{
-				ID:        fmt.Sprintf("%d", upload.ID),
-				FilePath:  upload.FileKey,
-				MaxTokens: 500,
-				Loader:    docL,
-				Metadata:  metadataText,
-			})
-			files = append(files, f)
-
-			content, err := s3L.GetFileText(ctx, f)
-			if err != nil {
-				return err
-			}
-
-			pdfBytes, err := loader.TransformDocToPdf(content, ext)
-			if err != nil {
-				return err
-			}
-			pages, err := loader.CountPDFPages(pdfBytes)
-			if err != nil {
-				return err
-			}
-			pageCount += pages
-		case "pptx":
-			pptxL := pptx.NewPPTXOcrGraphLoader(s3L, ocrL)
-			f := loader.NewGraphDocumentFile(loader.NewGraphFileParams{
-				ID:        fmt.Sprintf("%d", upload.ID),
-				FilePath:  upload.FileKey,
-				MaxTokens: 500,
-				Loader:    pptxL,
-				Metadata:  metadataText,
-			})
-			files = append(files, f)
-
-			content, err := s3L.GetFileText(ctx, f)
-			if err != nil {
-				return err
-			}
-
-			pdfBytes, err := loader.TransformDocToPdf(content, ext)
-			if err != nil {
-				return err
-			}
-			pages, err := loader.CountPDFPages(pdfBytes)
-			if err != nil {
-				return err
-			}
-			pageCount += pages
-		case "pdf":
-			ocrL := ocr.NewOCRGraphLoader(ocr.NewOCRGraphLoaderParams{
-				Loader:   nil,
-				AIClient: aiClient,
-			})
-			pdfL := pdf.NewPDFOcrGraphLoader(s3L, ocrL)
-			f := loader.NewGraphDocumentFile(loader.NewGraphFileParams{
-				ID:        fmt.Sprintf("%d", upload.ID),
-				FilePath:  upload.FileKey,
-				MaxTokens: 500,
-				Loader:    pdfL,
-				Metadata:  metadataText,
-			})
-			files = append(files, f)
-
-			content, err := s3L.GetFileText(ctx, f)
-			if err != nil {
-				return err
-			}
-
-			pages, err := loader.CountPDFPages(content)
-			if err != nil {
-				return err
-			}
-			pageCount += pages
-		case "jpg", "jpeg", "png", "bmp", "gif", "tiff", "heic", "webp":
-			imgL := image.NewImageGraphLoader(image.NewImageGraphLoaderParams{
-				AIClient: aiClient,
-				Loader:   s3L,
-			})
-			f := loader.NewGraphImageFile(loader.NewGraphFileParams{
-				ID:        fmt.Sprintf("%d", upload.ID),
-				FilePath:  upload.FileKey,
-				MaxTokens: 500,
-				Loader:    imgL,
-				Metadata:  metadataText,
-			})
-			files = append(files, f)
-			pageCount += 1
-		case "mp3", "wav", "mpeg", "mpga", "m4a", "ogg", "webm":
-			audioL := audio.NewAudioGraphLoader(audio.NewAudioGraphLoaderParams{
-				AIClient: aiClient,
-				Loader:   s3L,
-			})
-			f := loader.NewGraphAudioFile(loader.NewGraphFileParams{
-				ID:        fmt.Sprintf("%d", upload.ID),
-				FilePath:  upload.FileKey,
-				MaxTokens: 500,
-				Loader:    audioL,
-				Metadata:  metadataText,
-			})
-			files = append(files, f)
-
-			head, err := s3Client.HeadObject(ctx, &awss3.HeadObjectInput{
-				Bucket: aws.String(util.GetEnv("AWS_BUCKET")),
-				Key:    aws.String(upload.FileKey),
-			})
-			if err == nil && head.ContentLength != nil {
-				sizeMB := *head.ContentLength / (1024 * 1024)
-				if sizeMB < 1 {
-					pageCount += 1
-				} else {
-					pageCount += int(sizeMB)
-				}
-			} else {
-				pageCount += 1
-			}
-		case "csv":
-			csvL := csv.NewCSVGraphLoader(s3L)
-			f := loader.NewGraphCSVFile(loader.NewGraphFileParams{
-				ID:        fmt.Sprintf("%d", upload.ID),
-				FilePath:  upload.FileKey,
-				MaxTokens: 500,
-				Loader:    csvL,
-				Metadata:  metadataText,
-			})
-			files = append(files, f)
-
-			head, err := s3Client.HeadObject(ctx, &awss3.HeadObjectInput{
-				Bucket: aws.String(util.GetEnv("AWS_BUCKET")),
-				Key:    aws.String(upload.FileKey),
-			})
-			if err == nil && head.ContentLength != nil {
-				sizeKB := *head.ContentLength / 1024
-				pages := max(int(sizeKB/50), 1)
-				pageCount += pages
-			} else {
-				pageCount += 1
-			}
-		case "xlsx", "xls":
-			excelL := excel.NewExcelGraphLoader(s3L)
-
-			tempFile := loader.NewGraphDocumentFile(loader.NewGraphFileParams{
-				ID:        fmt.Sprintf("%d", upload.ID),
-				FilePath:  upload.FileKey,
-				MaxTokens: 500,
-				Loader:    s3L,
-				Metadata:  metadataText,
-			})
-			content, err := s3L.GetFileText(ctx, tempFile)
-			if err != nil {
-				return err
-			}
-
-			csvSheets, err := loader.TransformExcelToCsv(content, ext)
-			if err != nil {
-				return err
-			}
-
-			sheetIndex := 0
-			for sheetName, csvContent := range csvSheets {
-				parsed, parseErr := csv.ParseCSV(csvContent)
-				if parseErr != nil {
-					continue
-				}
-				if len(parsed) == 0 {
-					continue
-				}
-
-				sheetID := fmt.Sprintf("%d-sheet-%d", upload.ID, sheetIndex)
-				baseName := strings.TrimSuffix(filepath.Base(upload.FileKey), filepath.Ext(upload.FileKey))
-				sheetFilePath := fmt.Sprintf("%s/%s_%s%s",
-					filepath.Dir(upload.FileKey),
-					baseName,
-					sheetName,
-					filepath.Ext(upload.FileKey))
-
-				f := loader.NewGraphCSVFile(loader.NewGraphFileParams{
-					ID:        sheetID,
-					FilePath:  sheetFilePath,
-					MaxTokens: 500,
-					Loader:    excelL,
-					Metadata:  metadataText,
-				})
-				files = append(files, f)
-				sheetIndex++
-
-				sizeKB := len(parsed) / 1024
-				pages := max(sizeKB/2, 1)
-				pageCount += pages
-			}
-		case "txt", "md":
-			f := loader.NewGraphDocumentFile(loader.NewGraphFileParams{
-				ID:        fmt.Sprintf("%d", upload.ID),
-				FilePath:  upload.FileKey,
-				MaxTokens: 500,
-				Loader:    s3L,
-				Metadata:  metadataText,
-			})
-			files = append(files, f)
-
-			head, err := s3Client.HeadObject(ctx, &awss3.HeadObjectInput{
-				Bucket: aws.String(util.GetEnv("AWS_BUCKET")),
-				Key:    aws.String(upload.FileKey),
-			})
-			if err == nil && head.ContentLength != nil {
-				sizeKB := *head.ContentLength / 1024
-				pages := max(int(sizeKB/50), 1)
-				pageCount += pages
-			} else {
-				pageCount += 1
-			}
-		default:
-			f := loader.NewGraphGenericFile(
-				loader.NewGraphFileParams{
-					ID:        fmt.Sprintf("%d", upload.ID),
-					FilePath:  upload.FileKey,
-					MaxTokens: 500,
-					Loader:    s3L,
-				},
-				upload.Name,
-			)
-			noProcessingFiles = append(noProcessingFiles, f)
-		}
-	}
-
-	prediction, err := q.PredictProjectProcessTime(ctx, db.PredictProjectProcessTimeParams{
-		Duration: int64(pageCount),
-		StatType: "file_processing",
-	})
-	if err != nil {
-		prediction = 0
-	}
-	logger.Info("[Queue] Prediction for preprocessing", "batch_id", data.BatchID, "pages", pageCount, "time_ms", prediction)
-
-	if data.CorrelationID != "" {
-		_ = q.UpdateBatchEstimatedDuration(ctx, db.UpdateBatchEstimatedDurationParams{
-			CorrelationID:     data.CorrelationID,
-			BatchID:           int32(data.BatchID),
-			EstimatedDuration: prediction,
-		})
-	}
-
-	enc, err := tiktoken.GetEncoding("o200k_base")
-	if err != nil {
-		return err
-	}
-
-	type fileTokenCount struct {
-		fileID     int64
-		tokenCount int32
-		metadata   string
-		isGeneric  bool
-	}
-	tokenCounts := make([]fileTokenCount, 0)
-
-	start := time.Now()
-	for _, f := range files {
-		txt, err := f.GetText(ctx)
-		if err != nil {
-			return fmt.Errorf("get text for file %s: %w", f.ID, err)
-		}
-
-		textContent := string(txt)
-
-		fileID := f.ID
-		if idx := strings.Index(fileID, "-sheet-"); idx != -1 {
-			fileID = fileID[:idx]
-		}
-		id, err := strconv.ParseInt(fileID, 10, 64)
-		if err != nil {
-			return err
-		}
-
-		var fileName string
-		for _, upload := range *data.ProjectFiles {
-			if fmt.Sprintf("%d", upload.ID) == fileID {
-				fileName = upload.Name
-				break
-			}
-		}
-
-		metadata, err := ai.ExtractDocumentMetadata(ctx, aiClient, fileName, textContent)
-		if err != nil {
-			return fmt.Errorf("extract metadata for file %s: %w", f.ID, err)
-		}
-
-		cleanText := ai.StripMetadataTags(textContent)
-
-		tokens := enc.Encode(cleanText, nil, nil)
-		count := len(tokens)
-
-		tokenCounts = append(tokenCounts, fileTokenCount{fileID: id, tokenCount: int32(count), metadata: metadata, isGeneric: false})
-
-		key := filepath.Base(f.FilePath)
-		path := filepath.Dir(f.FilePath)
-		ext := filepath.Ext(key)
-		nameWithoutExt := strings.TrimSuffix(key, ext)
-		name := nameWithoutExt + ".txt"
-
-		content := bytes.NewReader([]byte(cleanText))
-		_, err = storage.PutFile(ctx, s3Client, path, name, nameWithoutExt, content)
-		if err != nil {
-			return fmt.Errorf("put file %s: %w", f.ID, err)
-		}
-	}
-	duration := time.Since(start)
-
-	for _, f := range noProcessingFiles {
-		content, err := f.GetText(ctx)
-		if err != nil {
-			return fmt.Errorf("get text for file %s: %w", f.ID, err)
-		}
-
-		textContent := string(content)
-
-		id, err := strconv.ParseInt(f.ID, 10, 64)
-		if err != nil {
-			return err
-		}
-
-		tokens := enc.Encode(textContent, nil, nil)
-		count := len(tokens)
-		tokenCounts = append(tokenCounts, fileTokenCount{fileID: id, tokenCount: int32(count), metadata: "", isGeneric: true})
-
-		if strings.TrimSpace(textContent) == "" {
-			continue
-		}
-
-		key := filepath.Base(f.FilePath)
-		path := filepath.Dir(f.FilePath)
-		ext := filepath.Ext(key)
-		nameWithoutExt := strings.TrimSuffix(key, ext)
-		name := nameWithoutExt + ".txt"
-
-		contentReader := bytes.NewReader([]byte(textContent))
-		_, err = storage.PutFile(ctx, s3Client, path, name, nameWithoutExt, contentReader)
-		if err != nil {
-			return fmt.Errorf("put file %s: %w", f.ID, err)
-		}
-	}
-
-	tx, err := conn.Begin(ctx)
-	if err != nil {
-		return err
-	}
-	defer tx.Rollback(ctx)
-	qtx := q.WithTx(tx)
-
-	for _, tc := range tokenCounts {
-		err = qtx.AddTokenCountToFile(ctx, db.AddTokenCountToFileParams{
-			ID:         tc.fileID,
-			TokenCount: tc.tokenCount,
-		})
-		if err != nil {
-			return fmt.Errorf("add token count for file %d: %w", tc.fileID, err)
-		}
-
-		if tc.isGeneric {
-			continue
-		}
-
-		err = qtx.UpdateProjectFileMetadata(ctx, db.UpdateProjectFileMetadataParams{
-			ID:       tc.fileID,
-			Metadata: pgtype.Text{String: tc.metadata, Valid: true},
-		})
-		if err != nil {
-			return fmt.Errorf("update metadata for file %d: %w", tc.fileID, err)
-		}
-	}
-
-	err = qtx.AddProcessTime(ctx, db.AddProcessTimeParams{
-		ProjectID: data.ProjectID,
-		Amount:    int32(pageCount),
-		Duration:  duration.Milliseconds(),
-		StatType:  "file_processing",
-	})
-	if err != nil {
-		return fmt.Errorf("add process time: %w", err)
-	}
-
-	err = tx.Commit(ctx)
-	if err != nil {
-		return err
-	}
-
-	if data.CorrelationID != "" {
-		_ = q.UpdateBatchStatus(ctx, db.UpdateBatchStatusParams{
-			CorrelationID: data.CorrelationID,
-			BatchID:       int32(data.BatchID),
-			Column3:       "preprocessed",
-		})
-	}
-
-	err = PublishFIFO(ch, "graph_queue", []byte(msg))
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
 
 func RecoverStaleBatches(
 	ctx context.Context,
@@ -1096,397 +150,4 @@ func ResetBatchStatusForRetry(
 			JobID:         int32(descMsg.JobID),
 		})
 	}
-}
-
-func KickoffDescriptionJobsForCorrelation(
-	ctx context.Context,
-	ch *amqp091.Channel,
-	pool *pgxpool.Pool,
-	correlationID string,
-	projectID int64,
-) (int, error) {
-	q := db.New(pool)
-
-	jobs, err := q.GetDescriptionJobsByCorrelation(ctx, correlationID)
-	if err != nil {
-		return 0, err
-	}
-
-	if len(jobs) == 0 {
-		batches, err := q.GetBatchesByCorrelation(ctx, correlationID)
-		if err != nil {
-			return 0, fmt.Errorf("failed to get batches: %w", err)
-		}
-
-		fileIDSet := make(map[int64]struct{})
-		for _, batch := range batches {
-			for _, fid := range batch.FileIds {
-				fileIDSet[fid] = struct{}{}
-			}
-		}
-
-		if len(fileIDSet) == 0 {
-			return 0, nil
-		}
-
-		fileIDs := make([]int64, 0, len(fileIDSet))
-		for fid := range fileIDSet {
-			fileIDs = append(fileIDs, fid)
-		}
-		slices.Sort(fileIDs)
-
-		entities, err := q.GetEntitiesWithSourcesFromFiles(ctx, db.GetEntitiesWithSourcesFromFilesParams{
-			Column1:   fileIDs,
-			ProjectID: projectID,
-		})
-		if err != nil {
-			return 0, fmt.Errorf("failed to get entities: %w", err)
-		}
-		rels, err := q.GetRelationshipsWithSourcesFromFiles(ctx, db.GetRelationshipsWithSourcesFromFilesParams{
-			Column1:   fileIDs,
-			ProjectID: projectID,
-		})
-		if err != nil {
-			return 0, fmt.Errorf("failed to get relationships: %w", err)
-		}
-
-		type item struct {
-			kind string
-			id   int64
-		}
-		items := make([]item, 0, len(entities)+len(rels))
-		for _, e := range entities {
-			items = append(items, item{kind: "entity", id: e.ID})
-		}
-		for _, r := range rels {
-			items = append(items, item{kind: "relationship", id: r.ID})
-		}
-		sort.Slice(items, func(i, j int) bool {
-			if items[i].kind == items[j].kind {
-				return items[i].id < items[j].id
-			}
-			return items[i].kind < items[j].kind
-		})
-
-		jobSize := int(util.GetEnvNumeric("AI_PARALLEL_REQ", 10))
-		totalJobs := (len(items) + jobSize - 1) / jobSize
-		if totalJobs == 0 {
-			return 0, nil
-		}
-
-		for i := range totalJobs {
-			start := i * jobSize
-			end := min(start+jobSize, len(items))
-			chunk := items[start:end]
-
-			entityIDs := make([]int64, 0, len(chunk))
-			relIDs := make([]int64, 0, len(chunk))
-			for _, it := range chunk {
-				switch it.kind {
-				case "entity":
-					entityIDs = append(entityIDs, it.id)
-				case "relationship":
-					relIDs = append(relIDs, it.id)
-				}
-			}
-
-			_, err := q.CreateDescriptionJobStatus(ctx, db.CreateDescriptionJobStatusParams{
-				ProjectID:       projectID,
-				CorrelationID:   correlationID,
-				JobID:           int32(i + 1),
-				TotalJobs:       int32(totalJobs),
-				EntityIds:       entityIDs,
-				RelationshipIds: relIDs,
-			})
-			if err != nil {
-				return 0, fmt.Errorf("failed to create description job status: %w", err)
-			}
-		}
-
-		jobs, err = q.GetDescriptionJobsByCorrelation(ctx, correlationID)
-		if err != nil {
-			return 0, err
-		}
-	}
-
-	for _, job := range jobs {
-		if job.Status != "pending" && job.Status != "failed" {
-			continue
-		}
-		msg := QueueDescriptionJobMsg{
-			ProjectID:       job.ProjectID,
-			CorrelationID:   job.CorrelationID,
-			JobID:           int(job.JobID),
-			TotalJobs:       int(job.TotalJobs),
-			EntityIDs:       job.EntityIds,
-			RelationshipIDs: job.RelationshipIds,
-		}
-		b, err := json.Marshal(msg)
-		if err != nil {
-			return len(jobs), err
-		}
-		if err := PublishFIFO(ch, "description_queue", b); err != nil {
-			return len(jobs), err
-		}
-	}
-
-	return len(jobs), nil
-}
-
-func ProcessDescriptionMessage(
-	ctx context.Context,
-	aiClient ai.GraphAIClient,
-	ch *amqp091.Channel,
-	pool *pgxpool.Pool,
-	msg string,
-) error {
-	_ = ch
-	data := new(QueueDescriptionJobMsg)
-	if err := json.Unmarshal([]byte(msg), data); err != nil {
-		return err
-	}
-	if data.CorrelationID == "" {
-		return fmt.Errorf("missing correlation_id")
-	}
-	if data.JobID <= 0 {
-		return fmt.Errorf("invalid job_id")
-	}
-
-	q := db.New(pool)
-	latestCorrelationID, latestErr := q.GetLatestCorrelationForProject(ctx, data.ProjectID)
-	if latestErr == nil && latestCorrelationID != "" && latestCorrelationID != data.CorrelationID {
-		_ = q.UpdateDescriptionJobStatus(ctx, db.UpdateDescriptionJobStatusParams{
-			CorrelationID: data.CorrelationID,
-			JobID:         int32(data.JobID),
-			Column3:       "completed",
-			ErrorMessage:  pgtype.Text{String: "skipped: superseded correlation", Valid: true},
-		})
-		return nil
-	}
-
-	_, err := q.TryStartDescriptionJob(ctx, db.TryStartDescriptionJobParams{
-		CorrelationID: data.CorrelationID,
-		JobID:         int32(data.JobID),
-	})
-	if err != nil {
-		if errors.Is(err, pgx.ErrNoRows) {
-			return nil
-		}
-		return err
-	}
-
-	storageClient, err := graphstorage.NewGraphDBStorageWithConnection(ctx, pool, aiClient, []string{})
-	if err != nil {
-		_ = q.UpdateDescriptionJobStatus(ctx, db.UpdateDescriptionJobStatusParams{
-			CorrelationID: data.CorrelationID,
-			JobID:         int32(data.JobID),
-			Column3:       "failed",
-			ErrorMessage:  pgtype.Text{String: err.Error(), Valid: true},
-		})
-		return err
-	}
-
-	batches, err := q.GetBatchesByCorrelation(ctx, data.CorrelationID)
-	if err != nil {
-		return err
-	}
-	fileIDSet := make(map[int64]struct{})
-	for _, b := range batches {
-		for _, fid := range b.FileIds {
-			fileIDSet[fid] = struct{}{}
-		}
-	}
-	fileIDs := make([]int64, 0, len(fileIDSet))
-	for fid := range fileIDSet {
-		fileIDs = append(fileIDs, fid)
-	}
-	slices.Sort(fileIDs)
-
-	if err := storageClient.UpdateEntityDescriptionsByIDsFromFiles(ctx, data.EntityIDs, fileIDs); err != nil {
-		_ = q.UpdateDescriptionJobStatus(ctx, db.UpdateDescriptionJobStatusParams{
-			CorrelationID: data.CorrelationID,
-			JobID:         int32(data.JobID),
-			Column3:       "failed",
-			ErrorMessage:  pgtype.Text{String: err.Error(), Valid: true},
-		})
-		return err
-	}
-	if err := storageClient.UpdateRelationshipDescriptionsByIDsFromFiles(ctx, data.RelationshipIDs, fileIDs); err != nil {
-		_ = q.UpdateDescriptionJobStatus(ctx, db.UpdateDescriptionJobStatusParams{
-			CorrelationID: data.CorrelationID,
-			JobID:         int32(data.JobID),
-			Column3:       "failed",
-			ErrorMessage:  pgtype.Text{String: err.Error(), Valid: true},
-		})
-		return err
-	}
-
-	_ = q.UpdateDescriptionJobStatus(ctx, db.UpdateDescriptionJobStatusParams{
-		CorrelationID: data.CorrelationID,
-		JobID:         int32(data.JobID),
-		Column3:       "completed",
-	})
-
-	allDone, err := q.AreAllDescriptionJobsCompleted(ctx, data.CorrelationID)
-	if err != nil {
-		return err
-	}
-	if !allDone {
-		return nil
-	}
-
-	batchesDone, err := q.AreAllBatchesCompleted(ctx, data.CorrelationID)
-	if err != nil {
-		return err
-	}
-	if !batchesDone {
-		return nil
-	}
-
-	latestCorrelationID, latestErr = q.GetLatestCorrelationForProject(ctx, data.ProjectID)
-	if latestErr != nil {
-		return nil
-	}
-	if latestCorrelationID != data.CorrelationID {
-		return nil
-	}
-
-	if _, err := q.UpdateProjectState(ctx, db.UpdateProjectStateParams{ID: data.ProjectID, State: "ready"}); err != nil {
-		logger.Error("[Queue] Failed to set project state to ready", "project_id", data.ProjectID, "correlation_id", data.CorrelationID, "err", err)
-	}
-
-	return nil
-}
-
-// FinalizeDescriptionsForCorrelation generates descriptions for all entities/relationships
-// that have new sources from the batches in the given correlation ID.
-func FinalizeDescriptionsForCorrelation(
-	ctx context.Context,
-	pool *pgxpool.Pool,
-	aiClient ai.GraphAIClient,
-	correlationID string,
-) error {
-	q := db.New(pool)
-
-	batches, err := q.GetBatchesByCorrelation(ctx, correlationID)
-	if err != nil {
-		return fmt.Errorf("failed to get batches: %w", err)
-	}
-
-	fileIDSet := make(map[int64]struct{})
-	for _, batch := range batches {
-		for _, fid := range batch.FileIds {
-			fileIDSet[fid] = struct{}{}
-		}
-	}
-
-	if len(fileIDSet) == 0 {
-		logger.Debug("[Queue] No files to generate descriptions for", "correlation_id", correlationID)
-		return nil
-	}
-
-	fileIDs := make([]int64, 0, len(fileIDSet))
-	for fid := range fileIDSet {
-		fileIDs = append(fileIDs, fid)
-	}
-
-	files := make([]loader.GraphFile, 0, len(fileIDs))
-	for _, fid := range fileIDs {
-		files = append(files, loader.GraphFile{ID: fmt.Sprintf("%d", fid)})
-	}
-
-	logger.Info("[Queue] Starting description generation with retry", "correlation_id", correlationID, "file_count", len(files))
-
-	const maxRetries = 3
-	const baseBackoffMs = 100
-
-	var lastErr error
-	for attempt := range maxRetries {
-		if attempt > 0 {
-			backoffDuration := time.Duration((attempt)*baseBackoffMs) * time.Millisecond
-			logger.Debug("[Queue] Retrying description generation", "correlation_id", correlationID, "attempt", attempt+1, "backoff_ms", backoffDuration.Milliseconds())
-
-			select {
-			case <-ctx.Done():
-				return ctx.Err()
-			case <-time.After(backoffDuration):
-			}
-		}
-
-		storageClient, err := graphstorage.NewGraphDBStorageWithConnection(ctx, pool, aiClient, []string{})
-		if err != nil {
-			lastErr = fmt.Errorf("failed to create storage client: %w", err)
-			logger.Warn("[Queue] Failed to create storage client", "correlation_id", correlationID, "attempt", attempt+1, "err", lastErr)
-			continue
-		}
-
-		err = storageClient.GenerateDescriptions(ctx, files)
-		if err == nil {
-			return nil
-		}
-
-		lastErr = err
-		logger.Warn("[Queue] Description generation failed", "correlation_id", correlationID, "attempt", attempt+1, "err", err)
-	}
-
-	return fmt.Errorf("failed to generate descriptions after %d retries: %w", maxRetries, lastErr)
-}
-
-func FinalizeDescriptionsForCorrelationWithLock(
-	ctx context.Context,
-	pool *pgxpool.Pool,
-	aiClient ai.GraphAIClient,
-	correlationID string,
-) error {
-	lockConn, err := pool.Acquire(ctx)
-	if err != nil {
-		return fmt.Errorf("failed to acquire lock connection: %w", err)
-	}
-	defer lockConn.Release()
-
-	lockID := correlationLockID(correlationID)
-	q := db.New(lockConn)
-	acquired, err := q.TryAcquireProjectLock(ctx, lockID)
-	if err != nil {
-		return fmt.Errorf("failed to acquire description lock: %w", err)
-	}
-	if !acquired {
-		logger.Debug("[Queue] Description generation already in progress", "correlation_id", correlationID)
-		return nil
-	}
-	defer func() {
-		if unlockErr := q.ReleaseProjectLock(ctx, lockID); unlockErr != nil {
-			errMsg := unlockErr.Error()
-			if strings.Contains(errMsg, "conn closed") {
-				logger.Debug("[Queue] Failed to release description lock (connection closed, lock auto-released)", "correlation_id", correlationID, "err", unlockErr)
-			} else {
-				logger.Error("[Queue] Failed to release description lock", "correlation_id", correlationID, "err", unlockErr)
-			}
-		}
-	}()
-
-	allDone, checkErr := q.AreAllBatchesCompleted(ctx, correlationID)
-	if checkErr != nil {
-		return fmt.Errorf("failed to re-check batch completion: %w", checkErr)
-	}
-	if !allDone {
-		logger.Debug("[Queue] Batches no longer complete, skipping description generation", "correlation_id", correlationID)
-		return nil
-	}
-
-	logger.Info("[Queue] All batches completed, starting description generation", "correlation_id", correlationID)
-
-	err = FinalizeDescriptionsForCorrelation(ctx, pool, aiClient, correlationID)
-	if err != nil {
-		return err
-	}
-
-	logger.Info("[Queue] Description generation completed", "correlation_id", correlationID)
-	return nil
-}
-
-func correlationLockID(correlationID string) int64 {
-	hasher := fnv.New64a()
-	_, _ = hasher.Write([]byte("correlation:" + correlationID))
-	return int64(hasher.Sum64())
 }
